# backend/schema/openapi.yaml
openapi: 3.0.3
info:
  title: Assured Express Logistics API
  version: 1.0.0
  description: Operations Command Center — Riders, Merchants, Zones, Communications
paths:
...
  /api/v1/core/orders/{id}/:
    get:
      operationId: v1_core_orders_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this order.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: ''
    put:
      operationId: v1_core_orders_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this order.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusUpdate'
          description: ''
    patch:
      operationId: v1_core_orders_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this order.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderStatusUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderStatusUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderStatusUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusUpdate'
          description: ''
    delete:
      operationId: v1_core_orders_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this order.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/core/orders/{id}/assign/:
    post:
      operationId: v1_core_orders_assign_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this order.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Order'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: ''
...
  /api/v1/core/riders/{id}/orders/:
    get:
      operationId: v1_core_riders_orders_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this rider.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderDetail'
          description: ''
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        reference:
          type: string
          maxLength: 50
        status:
          $ref: '#/components/schemas/StatusE5dEnum'
        merchant:
          type: integer
        merchant_name:
          type: string
          readOnly: true
        rider:
          type: integer
          nullable: true
        rider_name:
          type: string
          readOnly: true
        zone:
          type: integer
        zone_name:
          type: string
          readOnly: true
        pickup_address:
          type: string
        delivery_address:
          type: string
        delivery_fee:
          type: integer
          maximum: 2147483647
          minimum: 0
        order_value:
          type: integer
          maximum: 2147483647
          minimum: 0
          description: Value of goods
        km_distance:
          type: string
          format: decimal
          pattern: ^-?\d{0,4}(?:\.\d{0,2})?$
          nullable: true
        csat_score:
          type: integer
          maximum: 5
          minimum: 1
          nullable: true
        rejection_reason:
          type: string
          maxLength: 200
        failure_reason:
          type: string
          maxLength: 200
        created_at:
          type: string
          format: date-time
          readOnly: true
        assigned_at:
          type: string
          format: date-time
          nullable: true
        picked_up_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
      required:
      - created_at
      - delivery_address
      - id
      - merchant
      - merchant_name
      - pickup_address
      - reference
      - rider_name
      - zone
      - zone_name
...
    OrderStatusUpdate:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/StatusE5dEnum'
        rider:
          type: integer
          nullable: true
        rejection_reason:
          type: string
          maxLength: 200
        failure_reason:
          type: string
          maxLength: 200
        csat_score:
          type: integer
          maximum: 5
          minimum: 1
          nullable: true
        picked_up_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
...
    PatchedOrderStatusUpdate:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/StatusE5dEnum'
        rider:
          type: integer
          nullable: true
        rejection_reason:
          type: string
          maxLength: 200
        failure_reason:
          type: string
          maxLength: 200
        csat_score:
          type: integer
          maximum: 5
          minimum: 1
          nullable: true
        picked_up_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
...
    RiderDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        full_name:
          type: string
          readOnly: true
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 254
        zone:
          type: integer
        zone_name:
          type: string
          readOnly: true
        vertical_name:
          type: string
          readOnly: true
        status:
          $ref: '#/components/schemas/StatusB5bEnum'
        joined_at:
          type: string
          format: date
        bike_plate:
          type: string
          maxLength: 20
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - first_name
      - full_name
      - id
      - joined_at
      - last_name
      - phone
      - updated_at
      - vertical_name
      - zone
      - zone_name
...
    StatusB5bEnum:
      enum:
      - active
      - inactive
      - suspended
      type: string
      description: |-
        * `active` - Active
        * `inactive` - Inactive
        * `suspended` - Suspended
...
    StatusE5dEnum:
      enum:
      - pending
      - assigned
      - picked_up
      - delivered
      - failed
      - returned
      - cancelled
      type: string
      description: |-
        * `pending` - Pending
        * `assigned` - Assigned to Rider
        * `picked_up` - Picked Up
        * `delivered` - Delivered
        * `failed` - Failed
        * `returned` - Returned
        * `cancelled` - Cancelled
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT